+++ lib/k8s/client/runner/stream.ex
@@ -4,9 +4,9 @@ defmodule K8s.Client.Runner.Stream do
   """
 
   alias K8s.Client.Runner.Base
-  alias K8s.Operation
-  alias K8s.Conn
   alias K8s.Client.Runner.Stream.ListRequest
+  alias K8s.Conn
+  alias K8s.Operation
 
   @typedoc "List of items and pagination request"
   @type state_t :: {list(), ListRequest.t()}